added logic to order_product destroy action method that adds quantity from the order product back into the inventory of the product and handles in the order_product record is not actually an record contained in the current cart

+++ app/controllers/order_products_controller.rb
@@ -11,10 +11,17 @@ class OrderProductsController < ApplicationController
 
   def destroy
     order_product = OrderProduct.find_by(id: params[:id])
-    if order_product.destroy
-      flash[:success] = "Successfully removed from Cart"
+    product = order_product.product
+    if @current_cart.order_products.include?(order_product)
+      quantity = order_product.quantity
+      product.inventory += quantity
+      if product.save && order_product.destroy
+        flash[:success] = "Successfully removed from Cart"
+      else
+        flash[:alert] = "Unable to delete from Cart at this moment"
+      end
     else
-      flash[:alert] = "Unable to delete from Cart at this moment"
+      flash[:alert] = "That doesn't exist in your Cart"
     end
     redirect_to order_path(@current_cart.id)
   end
